+++ src/main/java/until/the/eternity/config/openapi/OpenApiFilters.java
@@ -1,0 +1,4 @@
+package until.the.eternity.config.openapi;
+
+public class OpenApiFilter {
+}
